refactor(core): keep runTests a coarse orchestrator
`runTests` had grown to hold the browser-only run path, mixed-run browser
subset planning, and the browser watch control plane inline, so the
orchestrator read as browser-mode detail rather than as the run flow.

Move that detail behind collaborators, leaving `runTests` with the coarse
flow only (split projects -> node executor init barrier -> plan -> drive
executors -> finalize):

- `browserOnlyRun.ts` — the browser-only run path, still separated by the
  cold-start gate (with zero node projects a `NodeExecutor` would boot an
  Rsbuild instance that resolves to empty `environments`).
- `browserRunPlanner.ts` — file-filter classification, the
  `modifyRstestConfig` discovery boot, and the executor/watch option bags.
- `browserWatchControls.ts` — fatal-signal exit, the single stdin
  shortcuts owner, and a `BrowserWatchSession` wrapper so the mixed watch
  loop fans reruns out without tracking watch handles itself.

`runBrowserModeTests` moves to `browserLoader.ts` next to
`loadBrowserExecutor` so the core<->browser load boundary is one file;
`globalSetupFailureOutcome` and `createCoverageProviderWithLog` move to
the modules that own them, the latter no longer taking a run context.

Along the way, dedupe what the extraction made visible: the CLI
file-filter predicates were duplicated verbatim between the run and list
paths (now shared in `utils/testFiles.ts`), the coverage banner had two
sources, and the browser executor run-option field list had three copies.
The watch-rerun path collectors move to `watchState.ts` so the node watch
loop no longer imports from a browser-named module.

No behavior change: 611 lines of orchestrator down from 1173, with the
two deliberate exceptions to executor isomorphism (browser-only fast path,
watch dual-drive) documented at their call sites and in `AGENTS.md`.
